% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_performPrimer3.R
\name{performPrimer3}
\alias{performPrimer3}
\title{Generate PCR primers (local primer3).}
\usage{
performPrimer3(
  input.sequence,
  target = NULL,
  exclude = NULL,
  product.sizes = c("50-150", "150-250"),
  primer.size = c(18, 20, 23),
  primer.tm = c(57, 59, 62),
  nPrimer = 3,
  internalOligo = FALSE,
  mismatching.library = NULL,
  primer3.args = NULL,
  primer3_core.binary = NULL,
  primer3_config = NULL,
  primer3_defaultsettings = NULL
)
}
\arguments{
\item{input.sequence}{(\link[Biostrings]{XStringSet}): A DNAString or RNAString object containing the sequence.}

\item{target}{(\link[Biostrings]{XStringSet} or numeric): A DNAString or RNAString object containing a sequence that must be present in the PCR product.
Alternatively, a numeric vector can be given to include that starting position + width, e.g. c(10, 15) to include position 10 to 25.}

\item{exclude}{(\link[Biostrings]{XStringSet} or numeric): A DNAString or RNAString object containing a sequence that must not be present in the PCR product.
Alternatively, a numeric vector can be given to exclude that starting position + width, e.g. c(10, 15) to exlude position 10 to 25.}

\item{product.sizes}{(character): A vector of size ranges. Primer3 first tries to pick primers in the first range. If that is not possible, it goes to the next range and tries again.}

\item{primer.size}{(character): Min. optimal and max. primer sizes. E.g. c(18, 20, 23).}

\item{primer.tm}{(character): Min. optimal and max. primer Tm. E.g. c(57, 59, 62).}

\item{nPrimer}{(numeric): Max. number of primers pairs to return.}

\item{internalOligo}{(logical): Also generate an internal oligo?}

\item{mismatching.library}{(character): Path to a primer mismatching library (provided by primer3). Leave NULL to ignore.}

\item{primer3.args}{(list): Additional arguments for primer3 settings. E.g. list('PRIMER_PICK_INTERNAL_OLIGO' = 0, 'PRIMER_EXPLAIN_FLAG' = 1).}

\item{primer3_core.binary}{(character): Optional location of the binary of primer3_core (if not found by Sys.which())}

\item{primer3_config}{(character): Path to primer3_config folder.}

\item{primer3_defaultsettings}{(character): Path to primer3 default settings file.}
}
\value{
(\link[tibble]{tibble}) Results of primer3 in a tibble.
}
\description{
Performs a local primer3 analysis to generate PCR primers based on a given DNA or RNA sequence. 
Sequences which \strong{must} be included in the final PCR product can be given via \emph{target}, this will generate flanking primers accordingly.

Please see the primer3 instructions on more detailed information on the primer3-specific settings.
In addition, the path to the primer3_config folder need to specified and (if not detected) also the primer3 executable.
}
\examples{
 \donttest{
 
 DNA.MDM2 <- Biostrings::DNAString('ATGTGCAATACCAACATGTCTGTACCTACTGATGGTGCTGTAACCACCTCACAGATT
 CCAGCTTCGGAACAAGAGACCCTGGTTCTTTTTTATCTTGGCCAGTATATTATGACTAAACGATTATATGATGAGAAGCAACAACATATTGTA')
 primer3.results <- performPrimer3(DNA.MDM2, target = Biostrings::DNAString('TGGTTCTTTTT'), nPrimer = 3)
 
	}
}
