% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_runSystemCommand.R
\name{runSystemCommand}
\alias{runSystemCommand}
\title{Perform a system command.}
\usage{
runSystemCommand(cmd)
}
\arguments{
\item{cmd}{(character): Command to execute.}
}
\value{
(list) List of exitcode, stdout and stderr.
}
\description{
Performs a system command and captures STDERR and STDOUT in a list.
If the exitStatus is not zero (0), then the command failed.
}
\examples{
 r <- runSystemCommand('ls -lh | head -n 5')

}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
